requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
mongoengine==0.23.0
motor==2.3.1
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.16.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
